+++ tests/run.py
@@ -128,6 +128,7 @@ NAVEGADOR = [
     ("facturaparto.js", ["facturaparto.js", "http://127.0.0.1:8402/"], {}),
     ("sobrelamesa.js", ["sobrelamesa.js", "http://127.0.0.1:8402/"], {}),
     ("tiempoui.js", ["tiempoui.js", "http://127.0.0.1:8402/"], {}),
+    ("ajustesap.js", ["ajustesap.js", "http://127.0.0.1:8402/"], {}),
     ("rolesui.js", ["rolesui.js", "http://127.0.0.1:8412/"], {}),
     ("gal.js", ["gal.js", "http://127.0.0.1:8300/"], {}),
     ("cruz.js", ["cruz.js", "http://127.0.0.1:8306/"], {}),